refactor: BioTcpHandler
